move info commands to a public plugin
